Address PR feedback from Tomas and Jared
